@@ -8,7 +8,7 @@ public class Node {
     final double lon;
     final double lat;
     final long hash;
-    Map<Integer, Integer> nodeTags;
+    Map<String, String> nodeTags;
     boolean waypart;
 
     Node(long hash, double lon, double lat, boolean waypart) {
@@ -27,7 +27,7 @@ public class Node {
         this.waypart = waypart;
     }
 
-    public Node(long id, long hash, double lon, double lat, boolean waypart, Map<Integer, Integer> tags) {
+    public Node(long id, long hash, double lon, double lat, boolean waypart, Map<String, String> tags) {
         this.id = id;
         this.hash = hash;
         this.lon = lon;
@@ -42,7 +42,7 @@ public class Node {
         }
     }
 
-    void addTag(int key, int value) {
+    void addTag(String key, String value) {
         if (nodeTags == null) {
             nodeTags = new HashMap<>();
         }
@@ -70,7 +70,7 @@ public class Node {
         return hash;
     }
 
-    Map<Integer, Integer> getTags() {
+    Map<String, String> getTags() {
         return nodeTags;
     }
 
