@@ -6,7 +6,6 @@ import java.io.File;
 import java.io.IOException;
 import java.util.stream.Stream;
 
-@Deprecated
 public class OSMPBFWriter {
 
     private final File outFile;
@@ -21,16 +20,16 @@ public class OSMPBFWriter {
         op.writePBFHeaders();
     }
 
-    void writeOSMPBFElements(StringTable stringtable, Long2ObjectOpenHashMap<Node> nodes, Long2ObjectOpenHashMap<Way> ways, Long2ObjectOpenHashMap<Relation> relations) throws IOException {
-        op.writePBFElements(stringtable, nodes.values().stream(), null, null);
-        op.writePBFElements(stringtable, null, ways.values().stream(), null);
-        op.writePBFElements(stringtable, null, null, relations.values().stream());
+    void writeOSMPBFElements(Long2ObjectOpenHashMap<Node> nodes, Long2ObjectOpenHashMap<Way> ways, Long2ObjectOpenHashMap<Relation> relations) throws IOException {
+        op.writePBFElements(nodes.values().stream(), null, null);
+        op.writePBFElements(null, ways.values().stream(), null);
+        op.writePBFElements(null, null, relations.values().stream());
     }
 
-    void writeOSMPBFElements(StringTable stringtable, Stream<Node> nodes, Stream<Way> ways, Stream<Relation> relations) throws IOException {
-        op.writePBFElements(stringtable, nodes, null, null);
-        op.writePBFElements(stringtable, null, ways, null);
-        op.writePBFElements(stringtable, null, null, relations);
+    void writeOSMPBFElements(Stream<Node> nodes, Stream<Way> ways, Stream<Relation> relations) throws IOException {
+        op.writePBFElements(nodes, null, null);
+        op.writePBFElements(null, ways, null);
+        op.writePBFElements(null, null, relations);
     }
 
     void closeOSMPBFFile() throws IOException {
